pyproject.toml
[project]
name = "skin-disease"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.3",
    "fpdf2>=2.8.7",
    "groq>=1.1.2",
    "httpx>=0.28.1",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "plotly>=6.6.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "rav>=0.4.3",
    "requests>=2.33.1",
    "sqlalchemy>=2.0.48",
    "streamlit>=1.56.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "uvicorn>=0.42.0",
]
